Optimized position of slots in floorplan L-form

+++ Presentation/PresentationLogic/LearningSpace/SpaceLayout/FloorPlans/Point.cs
@@ -5,6 +5,6 @@
 /// </summary>
 public struct Point
 {
-    public int X { get; init; }
+    public double X { get; init; }
     public int Y { get; init; }
 }
